+++ src/main/java/br/com/thiagoRDS/api_authors/modules/authors/usecases/ForgotMailPassword.java
@@ -12,7 +12,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsReposito
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.RecoveryTokensRepository;
 import br.com.thiagoRDS.api_authors.providers.CurrentContextProvider;
 import br.com.thiagoRDS.api_authors.providers.MailProvider.MailProvider;
-import br.com.thiagoRDS.api_authors.providers.dtos.SendMailDTO;
+import br.com.thiagoRDS.api_authors.providers.MailProvider.dtos.SendMailDTO;
 
 @Service
 public class ForgotMailPassword {
