+++ src/main/java/br/com/thiagoRDS/api_authors/providers/MailProvider/Impl/MailProviderImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 
 import br.com.thiagoRDS.api_authors.exceptions.InternalServerErrorException;
 import br.com.thiagoRDS.api_authors.providers.MailProvider.MailProvider;
-import br.com.thiagoRDS.api_authors.providers.dtos.SendMailDTO;
+import br.com.thiagoRDS.api_authors.providers.MailProvider.dtos.SendMailDTO;
 import jakarta.mail.internet.MimeMessage;
 
 @Service
