+++ src/main/java/br/com/thiagoRDS/api_authors/providers/MailProvider/MailProvider.java
@@ -1,6 +1,6 @@
 package br.com.thiagoRDS.api_authors.providers.MailProvider;
 
-import br.com.thiagoRDS.api_authors.providers.dtos.SendMailDTO;
+import br.com.thiagoRDS.api_authors.providers.MailProvider.dtos.SendMailDTO;
 
 public abstract class MailProvider {
   public abstract void sendMail(SendMailDTO data);
