+++ src/main/java/br/com/thiagoRDS/api_authors/providers/MailProvider/dtos/SendMailDTO.java
@@ -1,0 +1,7 @@
+package br.com.thiagoRDS.api_authors.providers.MailProvider.dtos;
+
+public record SendMailDTO(
+                String to,
+                String subject,
+                String text) {
+}
