--- src/main/java/br/com/thiagoRDS/api_authors/providers/dtos/SendMailDTO.java
@@ -1,7 +1,0 @@
-package br.com.thiagoRDS.api_authors.providers.dtos;
-
-public record SendMailDTO(
-        String to,
-        String subject,
-        String text) {
-}
